fix(AOT): stop narrowing AOT source lines to 16 bits (#5399)
QoreProgramLineLocation and the AOT wire format both carry 32-bit line numbers,
but the AOT location collectors still assigned through static_cast<int16_t>, so
any source line above 32767 was stored with its high bit set and read back
negative: line 33794 became -31742.  Qorus generates single source files well
over 32767 lines; the report was QorusRestApiHandlerV9.qc at lines 33795/33797.

Two visible effects, both reproduced against the pre-fix build:

- the runtime location table skips entries with a non-positive start line, so an
  exception raised in such a file reported line -1
- the source-stripped statement-location table fed the negative value straight
  into QoreProgramLocation, whose `assert(sline >= -1)` aborted the process in a
  debug build (QoreLibIntern.h:380) -- the reported symptom.  That table is only
  materialized under PO_ALLOW_DEBUGGER, which is why it took debugger support to
  surface

The fix:

- remove all nine narrowing casts in the statement-location, outlined-helper,
  closure, function and method location collectors; the destination fields were
  already int32_t, so nothing else had to change
- legacy (pre-v11) u16 blobs now zero-extend, which recovers a line in
  32768-65535 written by the old truncating writer exactly (the narrowing was
  mod 65536); sign-extending turned those same lines negative.  0xffff still
  reads back as -1 (unknown) for the signed call sites
- every deserialized line that reaches a QoreProgramLocation is validated, so an
  artifact already built with the truncating writer degrades to "no location"
  instead of aborting on the assertion, and a non-positive statement start line
  is skipped rather than registered

Tests: examples/test/ir/AOTLargeSourceLineLocations.qtest pins the reported line
for AST, JIT, AOT and source-stripped AOT with debugger support (the
configuration that aborted), at the 32767/32768 and 65535/65536 boundaries as
well as at line 33794, in both AOT debug-info modes.  It fails against the
pre-fix build (3 of 4 cases).

Tested: full test suite 1057/1057; valgrind clean on the AOT compile and load
paths, including the >32767-line source-stripped artifact.

+++ lib/QoreAOT.cpp
@@ -1447,8 +1447,8 @@ static bool buildAOTProgramStatementLocIndex(QoreProgram* pgm,
         }
 
         AOTCompiledFuncWithSlots::AOTStmtLocEntry entry;
-        entry.start_line = static_cast<int16_t>(loc->start_line);
-        entry.end_line = static_cast<int16_t>(loc->end_line);
+        entry.start_line = loc->start_line;
+        entry.end_line = loc->end_line;
         entry.offset = loc->offset;
         entry.file = file;
         entry.source = source;
@@ -1513,8 +1513,8 @@ static bool attachAOTProgramStatementLocs(QoreProgram* pgm,
         }
 
         AOTCompiledFuncWithSlots::AOTStmtLocEntry entry;
-        entry.start_line = static_cast<int16_t>(loc->start_line);
-        entry.end_line = static_cast<int16_t>(loc->end_line);
+        entry.start_line = loc->start_line;
+        entry.end_line = loc->end_line;
         entry.offset = loc->offset;
         entry.file = file;
         entry.source = source;
@@ -4633,8 +4633,8 @@ static bool aotLowerOutlinedFnHelpers(std::vector<AOTOutlinedHelper>& helpers,
         }
         for (const auto& loc : lowerer.getAOTLocTable()) {
             AOTCompiledFuncWithSlots::AOTLocEntry entry;
-            entry.start_line = static_cast<int16_t>(loc.start_line);
-            entry.end_line = static_cast<int16_t>(loc.end_line);
+            entry.start_line = loc.start_line;
+            entry.end_line = loc.end_line;
             entry.file = loc.file;
             aot_locs.push_back(std::move(entry));
         }
@@ -17587,8 +17587,8 @@ static bool compileAOTClosureBodiesForOwner(size_t owner_index, QoreProgram* pgm
                     }
                     const auto& loc = loc_table[loc_index];
                     AOTCompiledFuncWithSlots::AOTLocEntry entry;
-                    entry.start_line = static_cast<int16_t>(loc.start_line);
-                    entry.end_line = static_cast<int16_t>(loc.end_line);
+                    entry.start_line = loc.start_line;
+                    entry.end_line = loc.end_line;
                     entry.file = loc.file;
                     aot_locs.push_back(std::move(entry));
                 }
@@ -19937,8 +19937,8 @@ static void compileNamespaceFunctions(qore_ns_private* ns, QoreProgram* pgm,
                         if (std_entry_ok) {
                             for (const auto& loc : retry_lowerer.getAOTLocTable()) {
                                 AOTCompiledFuncWithSlots::AOTLocEntry entry;
-                                entry.start_line = static_cast<int16_t>(loc.start_line);
-                                entry.end_line = static_cast<int16_t>(loc.end_line);
+                                entry.start_line = loc.start_line;
+                                entry.end_line = loc.end_line;
                                 entry.file = loc.file;
                                 aot_locs_local.push_back(std::move(entry));
                             }
@@ -20032,8 +20032,8 @@ static void compileNamespaceFunctions(qore_ns_private* ns, QoreProgram* pgm,
                     if (std_entry_ok && !used_retry_lowering) {
                         for (const auto& loc : lowerer.getAOTLocTable()) {
                             AOTCompiledFuncWithSlots::AOTLocEntry entry;
-                            entry.start_line = static_cast<int16_t>(loc.start_line);
-                            entry.end_line = static_cast<int16_t>(loc.end_line);
+                            entry.start_line = loc.start_line;
+                            entry.end_line = loc.end_line;
                             entry.file = loc.file;
                             aot_locs_local.push_back(std::move(entry));
                         }
@@ -20449,8 +20449,8 @@ static void compileNamespaceFunctions(qore_ns_private* ns, QoreProgram* pgm,
                             if (method_ok) {
                                 for (const auto& loc : retry_lowerer.getAOTLocTable()) {
                                     AOTCompiledFuncWithSlots::AOTLocEntry entry;
-                                    entry.start_line = static_cast<int16_t>(loc.start_line);
-                                    entry.end_line = static_cast<int16_t>(loc.end_line);
+                                    entry.start_line = loc.start_line;
+                                    entry.end_line = loc.end_line;
                                     entry.file = loc.file;
                                     aot_locs_local.push_back(std::move(entry));
                                 }
@@ -20533,8 +20533,8 @@ static void compileNamespaceFunctions(qore_ns_private* ns, QoreProgram* pgm,
                         if (method_ok && !used_retry_lowering) {
                             for (const auto& loc : lowerer.getAOTLocTable()) {
                                 AOTCompiledFuncWithSlots::AOTLocEntry entry;
-                                entry.start_line = static_cast<int16_t>(loc.start_line);
-                                entry.end_line = static_cast<int16_t>(loc.end_line);
+                                entry.start_line = loc.start_line;
+                                entry.end_line = loc.end_line;
                                 entry.file = loc.file;
                                 aot_locs_local.push_back(std::move(entry));
                             }
